chore: apply requested changes
